# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto - OpenAPI 3.0
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! \
    Directly from your PC, you can upload your photos, and they will be visible to everyone who is following you.
    
    Functional design specification:
    
    - The login is performed by specifying the username and it returns a crafted userID(token).
    - Once in, the user sees on the web app the stream of photos of the users that he follows ordered in reverse chronological order.
    - Each photo has information about the date-time it was uploaded, the number of likes and the number of comments.
    - Users can place a like to photos from other users(not from themselves);
    - Users can remove likes from photos of other users;
    - Users can add comments to any photo and can edit comments to any photo(even for themselves);
    - Only the author can remove comments from a photo;
    - Users have their own personal profiles.
    
    When you access to the personal profile of a user you can see:
    - user’s photos (in reverse chronological order), 
    - their followers,
    - their following, 
    - how many photos have been uploaded.
    
    User action:
    - change his/her own username,
    - follow/unfollow other users,
    - upload photos, 
    - remove photos: also remove likes and comments.
    - Users can ban other users.
...
    Alice can decide to remove the ban at any moment.
    
...
    
...
    
...
servers:
...
components:
  parameters:
    user_id:
      description: User whose profile belongs to. 
      name: user_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UserID" }
    photo_id:
      description: ID of the picture considered
      name: photo_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PhotoID" }
    comment_id:
      description: Unique comment identifier
      name: comment_id
      in: path
      schema: { $ref: "#/components/schemas/CommentID" }
      required: true
        
  # 1) Define the security scheme type (HTTP bearer)
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Profile:
      description: Represents the information seen in the Profile Page of a user
      type: object
      properties:
        userId: {$ref: "#/components/schemas/UserID"}
        username: {$ref: "#/components/schemas/Username"}
        pictures_count:
          description: Number of photos in the profile of the user
          type: integer
          example: 169
        followers_count:
          description: Number of users that follow the profile
          type: integer
          example: 2567
        follows_count:
          description: number of users that the user follows
          type: integer
          example: 689
        profile_picture_url:
          description: URL of the profile picture. Accepting only ID attached with.png/.jpg/.jpeg extensions.
          type: string
          pattern: '^[a-zA-Z0-9-_]*\.(png|jpg|jpeg)$'
          example: "Cksvy-ZXtXRE.jpg"
          minLength: 0
          maxLength: 150
        bio:
          description: Biography of the profile. |
            I'm keeping the pattern simple. Not doing crazy things. Just allowing alphanumeric characters and basic punctuation.
          type: string
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          example: "Discover what's next"
          # on WASAPhoto
          minLength: 0
          maxLength: 150
...
    Comment:
      description: It represents a comment object
      type: object
      properties:
        commentId: {$ref: "#/components/schemas/CommentID"}
        created_in:
          description: When the comment was written
          type: string
          format: date-time
          example: 2022-11-04T16:35:22-04:00
          minLength: 20
          maxLength: 25
        author: {$ref: "#/components/schemas/Username"} # Uniquely identifies the author of a comment
        body:
          description: |
            Text of the comment written. Preventing empty comments.
          type: string
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          minLength: 1
          maxLength: 300
          example: "Buongiorno Beppe!"
        modified_in:
          description: When the comment was modified
          type: string
          format: date-time
          example: 2022-11-04T19:35:22-04:00
          minLength: 20
          maxLength: 25
        is_reply_comment:
          description: States if the comment is a reply or not
          type: boolean
          default: false
          example: true
        parentId: {$ref: "#/components/schemas/CommentID"}
    UserID:
      description: |
        The simplified login tell us that, since there is no password, the username must be unique among the application. 
        Logging in with an existing name will grant access to that user information. 
        Hence it would make sense to use the username as ID since we don't have to represent real people but virtual people, namely usernames, by simplifying in this way the ReST OpenAPI. 
        However, since it's not applicable in a real life scenario, I will use a userID that's different from the username. Thus, I'm choosing 20 chars long strings because they can store several billions of information(64^20, since the alphabet is "[a-z][A-Z][0-9]-_", 26+26+10+2=64), they are nicer to read then integers since they are shorter and almost equally fast to compute.
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "U3dhZ2dlciByb2Nrcw"
      readOnly: true
    PhotoID:
      description: Unique photo identifier
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "CksvyZXtXRE"
      readOnly: true
    CommentID:
      description: Unique comment identifier.
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      example: "CksvyZXtXRE"
      minLength: 1
      maxLength: 20
      readOnly: true
    Username:
      description: It represents the name of the user in WASA Photo
      type: string
      example: LeoMessi88
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 3
      maxLength: 16
  responses:
...
    InternalServerError: 
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    UnauthorizedError:
      description: Access token is missing or invalid.
paths:
...
  /users/{user_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
    
    patch:
      tags:
        - User
      summary: Update username
      description: This can only be done by the logged in user.
      operationId: setMyUserName
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Username'
      responses:
        "200":
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Username'
        "204":
          description: No content. Update occured but didn't send back any representation or description.
        "400": { description: Invalid username supplied }
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "500": { $ref: "#/components/responses/InternalServerError" }
      security:
        - bearerAuth: []
    
    put:
      tags:
        - User
      summary: Update profile
      description: This can only be done by the logged in user.
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        "200":
          description: Profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "204":
          description: No content. Update occured but didn't send back any representation or description.
        "400": { description: Invalid profile supplied }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []

    delete:
      tags:
        - User
      summary: Delete my profile
      description: This can only be done by the logged in user.
      operationId: deleteUserProfile
      responses:
        "204": { description: User deleted successfully }
        "400": { description: Invalid username supplied }
        '401': { $ref: '#/components/responses/UnauthorizedError'}
        "404": { description: User not found }
      security:
        - bearerAuth: []
...
  /users/{user_id}/bans/{ban_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - name: ban_id
        in: path
        required: true
        description: User who is banning
        schema:
          { $ref: "#/components/schemas/UserID" }

    put:
      tags:
        - User
      summary: Ban user
      description: >
        If Alice bans Bob, Bob won't be able to see any information about Alice. Alice can decide to remove the ban at any moment. I'm using PUT for its idempotency.
      operationId: banUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserID'
      responses:
        "201":
          description: Ban created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserID"
        "204": {description: User already banned.}
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": {description: User can't ban himself.}
      security:
          - bearerAuth: []

    delete:
      tags:
        - User
      summary: Unban user
      description: This can only be done by the logged in user. Bob can now see information about Alice.
      operationId: unbanUser
      responses:
        "204": { description: User unbanned successfully. }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't unban himself. }
      security:
        - bearerAuth: []
...
  /photos/{photo_id}/comments/:
    parameters:
      - $ref: "#/components/parameters/photo_id"
      
    get:
      tags:
        - Comments
      summary: Read all comments
      description: Read all comments under the photo photoID.
      operationId: readCommentsPhoto
      responses: 
        "200":
          description: List of comments
          content:
            application/json: 
              schema:
                description: |
                  Represents the comments under a photo added by the users.
                  Users can see comments in batches of 50.
                type: array
                minItems: 0
                maxItems: 50
                items: { $ref: "#/components/schemas/Comment" }
      security:
        - bearerAuth: []
      
        
    post:
      tags:
        - Comments
      summary: Write a comment
      description: Write a comment on a photo of a user.
      operationId: commentPhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        "201": 
          description: Comment created
          content:
            application/json: 
              schema:
                { $ref: "#/components/schemas/Comment" } 
        "204":
          description: No content.
        "400":
          description: Bad request. Invalid input data.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /comments/{comment_id}:
    parameters:
      - $ref: "#/components/parameters/comment_id"

    patch:
       tags:
         - Comments
       summary: Modify comment
       description: |
         An user can modify its own comment(and only an own comment)
       operationId: modifyComment
       requestBody:
         content:
           application/json:
             schema:
              description: When writing a comment, the front sends the author and the body of the comment.
              properties:
                  author: {$ref: "#/components/schemas/Username"} # Uniquely identifies the author of a comment
                  body:
                    description: |
                      Text of the comment written. Preventing empty comments.
                    type: string
                    pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
                    minLength: 1
                    maxLength: 300
                    example: "Buongiorno Beppe!"
       responses:
         "200":
           description: Comment updated successfully.
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/Comment'
         "201": 
           description: Comment created
           content:
             application/json: 
               schema:
                 { $ref: "#/components/schemas/Comment" }
         "400":
           description: Bad request. Invalid input data.
         '401':
           $ref: '#/components/responses/UnauthorizedError'

       security:
         - bearerAuth: []
      
    delete:
      tags:
        - Comments
      summary: Delete a comment
      description: |
        The Author and the Owner of the profile can delete a specific comment under a specific photo.
      operationId: uncommentPhoto
      responses:
        "204":
          description: No content. Comment deleted.
        "400":
          description: Comment not Found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []